refactor(phase4): extract resource_struct_tokens from codegen.rs
Move `resource_struct_tokens`, garde/typed-object helpers, scalar-type
token helpers, and `list_query_tokens` (~1,294 lines) into
`codegen/resource_struct_tokens.rs`. Cross-module calls to parent
functions (`list_query_ident`, `create_payload_fields`, etc.) use the
`super::` path since child modules can access private parent items.
All 233 passing tests remain green; 19 pre-existing failures unchanged.
